requirements.txt
discord.py
pyyaml
pillow
